Refactor script execution and verification loops
This commit refactors the script execution and verification loops in the `run.sh` and `verify.sh` files.

In the `run.sh` file, a new feature has been added to display the current script being executed out of the total number of scripts.

Similarly, in the `verify.sh` file, a new feature has been added to display the current file being processed out of the total number of files.